backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
click==8.3.1
colorama==0.4.6
fastapi==0.135.2
h11==0.16.0
idna==3.11
neo4j==6.1.0
psycopg2-binary==2.9.11
PyJWT==2.10.1
pydantic==2.12.5
pydantic_core==2.41.5
pytz==2026.1.post1
starlette==1.0.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.42.0
